+++ src/main/java/de/wwu/scdh/oxbytei/commons/OxygenSelectionDialog.java
@@ -18,11 +18,13 @@ import ro.sync.ecss.extensions.api.AuthorAccess;
 import ro.sync.exml.workspace.api.PluginWorkspaceProvider;
 import ro.sync.exml.workspace.api.PluginWorkspace;
 
+import de.wwu.scdh.teilsp.exceptions.ConfigurationException;
 import de.wwu.scdh.teilsp.services.extensions.ILabelledEntriesProvider;
 import de.wwu.scdh.teilsp.services.extensions.LabelledEntry;
 import de.wwu.scdh.teilsp.services.extensions.ExtensionException;
 import de.wwu.scdh.teilsp.exceptions.UIException;
 import de.wwu.scdh.teilsp.ui.ISelectionDialog;
+import de.wwu.scdh.teilsp.services.extensions.ArgumentDescriptor;
 
 
 public class OxygenSelectionDialog
@@ -33,18 +35,32 @@ public class OxygenSelectionDialog
     boolean multiple;
     List<String> currentValue, result;
     List<ILabelledEntriesProvider> providers;
+    Map<String, String> arguments;
 
     public OxygenSelectionDialog(Frame frame) {}
 
     public OxygenSelectionDialog() {
     }
 
-    public void init(Map<String, String> arguments) {
-	if (arguments.containsKey("title")) {
-	    title = arguments.get("title");
-	} else {
-	    title = "Select";
-	}
+    private static final ArgumentDescriptor<String> ARGUMENT_TITLE =
+	ISelectionDialog.ARGUMENT_TITLE;
+
+    private static final ArgumentDescriptor<?>[] ARGUMENTS = new ArgumentDescriptor<?>[] {
+	ARGUMENT_TITLE
+	    };
+	
+    public ArgumentDescriptor<?>[] getArgumentDescriptor() {
+	return ARGUMENTS;
+    }	
+
+    public void init(Map<String, String> arguments)
+	throws ConfigurationException {
+	this.arguments = arguments;
+	title = ARGUMENT_TITLE.getValue(arguments);
+    }
+
+    public Map<String, String> getArguments() {
+	return arguments;
     }
 
     public void setup(List<String> currentVal,