+++ src/main/java/de/wwu/scdh/teilsp/extensions/LabelledEntriesFromConfig.java
@@ -8,9 +8,11 @@ import javax.xml.transform.URIResolver;
 import org.xml.sax.EntityResolver;
 import org.w3c.dom.Document;
 
+import de.wwu.scdh.teilsp.exceptions.ConfigurationException;
 import de.wwu.scdh.teilsp.services.extensions.ILabelledEntriesProvider;
 import de.wwu.scdh.teilsp.services.extensions.ExtensionException;
 import de.wwu.scdh.teilsp.services.extensions.ArgumentDescriptor;
+import de.wwu.scdh.teilsp.services.extensions.ArgumentDescriptorImpl;
 import de.wwu.scdh.teilsp.services.extensions.LabelledEntry;
 
 
@@ -20,53 +22,42 @@ public class LabelledEntriesFromConfig
     private Map<String, String> arguments;
 
     private List<LabelledEntry> labelledEntries;
-    
-    private static final ArgumentDescriptor ARGUMENT_KEYS =
-	new ArgumentDescriptor("keys",
-			       ArgumentDescriptor.TYPE_STRING,
-			       "A list of keys.");
 
-    private static final ArgumentDescriptor ARGUMENT_LABELS =
-	new ArgumentDescriptor("labels",
-			       ArgumentDescriptor.TYPE_STRING,
-			       "A list of labels.");
+    private static final ArgumentDescriptor<String> ARGUMENT_KEYS =
+	new ArgumentDescriptorImpl<String>(String.class, "keys", "A list of keys.");
 
-    private static final ArgumentDescriptor ARGUMENT_SEPARATOR =
-	new ArgumentDescriptor("separator",
-			       ArgumentDescriptor.TYPE_STRING,
-			       "The separator that delimits the keys and labels. Default to comma.",
-			       ",");
+    private static final ArgumentDescriptor<String> ARGUMENT_LABELS =
+	new ArgumentDescriptorImpl<String>(String.class, "labels", "A list of labels.");
 
+    private static final ArgumentDescriptor<String> ARGUMENT_SEPARATOR =
+	new ArgumentDescriptorImpl<String>
+	(String.class, "separator",
+	 "The separator that delimits the keys and labels. Default to comma.", ",");
 
-    
     /**
      * The array of arguments, this author operation takes.
      */
-    private static final ArgumentDescriptor[] ARGUMENTS = new ArgumentDescriptor[] {
+    private static final ArgumentDescriptor<?>[] ARGUMENTS = new ArgumentDescriptor<?>[] {
 	ARGUMENT_KEYS,
 	ARGUMENT_LABELS,
 	ARGUMENT_SEPARATOR
 	};
 
     /**
-     * 
+     * See {@link ConfigurablePlugin#getArgumentDescriptor()}
      */
-    public ArgumentDescriptor[] getArgumentDescriptor() {
+    public ArgumentDescriptor<?>[] getArgumentDescriptor() {
 	return ARGUMENTS;
     }
 
-    public void init(Map<String, String> args,
-		     URIResolver uriResolver,
-		     EntityResolver entityResolver,
-		     Document doc,
-		     String systemId)
-	throws ExtensionException {
+    public void init(Map<String, String> args)
+	throws ConfigurationException {
 
 	arguments = args;
 
-	String sep = args.get("separator");
-	String[] keys = args.get("keys").split(sep);
-	String[] labels = args.get("labels").split(sep);
+	String sep = ARGUMENT_SEPARATOR.getValue(args);
+	String[] keys = ARGUMENT_KEYS.getValue(args).split(sep);
+	String[] labels = ARGUMENT_LABELS.getValue(args).split(sep);
 
 	labelledEntries = new ArrayList<LabelledEntry>();
 
@@ -74,7 +65,14 @@ public class LabelledEntriesFromConfig
 	for (int i = 0; i < l; i++) {
 	    labelledEntries.add(new LabelledEntry(keys[i].trim(), labels[i].trim()));
 	}
+    }
 
+    public void setup
+	(URIResolver uriResolver,
+	 EntityResolver entityResolver,
+	 Document doc,
+	 String systemId,
+	 String context) {
     }
 
     public Map<String, String> getArguments() {
@@ -85,5 +83,4 @@ public class LabelledEntriesFromConfig
 	throws ExtensionException {
 	return labelledEntries;
     }
-    
 }