+++ src/main/java/de/wwu/scdh/teilsp/ui/ComboBoxSelectDialog.java
@@ -37,6 +37,8 @@ import javax.swing.SwingConstants;
 import de.wwu.scdh.teilsp.services.extensions.ILabelledEntriesProvider;
 import de.wwu.scdh.teilsp.services.extensions.LabelledEntry;
 import de.wwu.scdh.teilsp.services.extensions.ExtensionException;
+import de.wwu.scdh.teilsp.services.extensions.ArgumentDescriptor;
+
 
 
 /**
@@ -56,6 +58,7 @@ public class ComboBoxSelectDialog
     JComboBox<LabelledEntry> comboBox;
     JLabel label;
     JPanel comboBoxes;
+    Map<String, String> arguments;
 
     public ComboBoxSelectDialog() {
 	super();
@@ -65,7 +68,27 @@ public class ComboBoxSelectDialog
 	super(frame, true);
     }
 
+    private static final ArgumentDescriptor<String> ARGUMENT_TITLE =
+	ISelectionDialog.ARGUMENT_TITLE;
+
+    private static final ArgumentDescriptor<URL> ARGUMENT_ICON =
+	ISelectionDialog.ARGUMENT_ICON;
+
+    public Map<String, String> getArguments() {
+	return arguments;
+    }
+
+    private static final ArgumentDescriptor<?>[] ARGUMENTS = new ArgumentDescriptor<?>[] {
+	ARGUMENT_TITLE,
+	    ARGUMENT_ICON
+	    };
+    
+    public ArgumentDescriptor<?>[] getArgumentDescriptor() {
+	return ARGUMENTS;
+    }	
+
     public void init(Map<String, String> arguments) {
+	this.arguments = arguments;
 	String title;
 	if (arguments.containsKey("title")) {
 	    title = arguments.get("title");